Bluetooth Bannana Sandwich
-OK so this think is sorta crazy and I'll try to explain it here.
First, the base here is emui5 and the ROM is Oreo and the
bluetooth default interface (default treble hal and bluetooth.default)
in the AOSP Oreo base just do not work on this device afaict. So
this solution was born of the idea to get the Nougat system
working on Oreo. It turns out there is no one big thing prevent
that senario from working. Just a lot of little things, the
most important being, the main bt interface definition having
been change to remove one function from the middle of the
stack.

-libhardware-nougat-bt-on-oreo: puts the missing function
back in to the stack in its place allowing the functions
below to line up correctly.

-bluetooth-crappy-stop-crash: dirty stuff to stop crashes
in libbluetooth_jni.

-Bluetooth-add-extra-interface-member-for-building: add
missing fuction to Oreo bt so build does not fail.

-bluetooth.default.so is from Nougat. Its' built with
some hacks to prevent crashing. It would be better to
bring it up from Nougat into Oreo but thats a huge mess.

-bluetooth.default.so also needs some shims defined in
the coh vendor repo.

-IMHO, it's extrodinary that this works as well as it does. The
BT interface definition in libhardware/include has
changed considerably and their are probably more mis-aligned
stacks to spot..